Calculate and show total value at end of entry

+++ src/database/entries_serialisation.go
@@ -14,7 +14,7 @@ func (er *DecimalValue) Scan(value any) error {
 		}
 
 		er.Whole = int64(binary.LittleEndian.Uint64(value))
-		er.Fractional = binary.LittleEndian.Uint64(value[8:])
+		er.Fractional = uint8(binary.LittleEndian.Uint16(value[8:]))
 
 	default:
 		return fmt.Errorf("UNMARSHALLING INVALID `DecimalValue`: %+v", value)
@@ -27,7 +27,7 @@ func (er DecimalValue) Value() (driver.Value, error) {
 	result := make([]byte, 16)
 
 	binary.LittleEndian.PutUint64(result, uint64(er.Whole))
-	binary.LittleEndian.PutUint64(result[8:], er.Fractional)
+	binary.LittleEndian.PutUint16(result[8:], uint16(er.Fractional))
 
 	return result, nil
 }
